Cleanup
Reorganized to prevent unrelated code from being in the same file.
