+++ Sources/ParseSwift/Object Protocols/ObjectType+Batch.swift
