+++ Sources/ParseSwift/Utilities/Encoding.swift
@@ -1,0 +1,39 @@
+//
+//  Encoding.swift
+//  ParseSwift (iOS)
+//
+//  Created by Pranjal Satija on 9/10/17.
+//  Copyright © 2017 Parse. All rights reserved.
+//
+
+import Foundation
+
+private let forbiddenKeys = ["createdAt", "updatedAt", "objectId", "className"]
+
+func getJSONEncoder() -> JSONEncoder {
+    let encoder = JSONEncoder()
+    encoder.dateEncodingStrategy = dateEncodingStrategy
+    return encoder
+}
+
+func getParseEncoder() -> ParseEncoder {
+    let encoder = ParseEncoder()
+    encoder.dateEncodingStrategy = parseDateEncodingStrategy
+    encoder.shouldEncodeKey = { (key, path) -> Bool in
+        if path.count == 0 // top level
+            && forbiddenKeys.index(of: key) != nil {
+            return false
+        }
+        return true
+    }
+    return encoder
+}
+
+extension JSONEncoder {
+    func encodeAsString<T>(_ value: T) throws -> String where T: Encodable {
+        guard let string = String(data: try encode(value), encoding: .utf8) else {
+            throw ParseError(code: -1, error: "Unable to encode object...")
+        }
+        return string
+    }
+}
