+++ Sources/ParseSwift/Utilities/NoBody.swift
@@ -1,0 +1,11 @@
+//
+//  NoBody.swift
+//  ParseSwift (iOS)
+//
+//  Created by Pranjal Satija on 9/10/17.
+//  Copyright © 2017 Parse. All rights reserved.
+//
+
+import Foundation
+
+public struct NoBody: Codable {}
